bluesn0w

+++ commands/minecraft.js
@@ -4,65 +4,32 @@ const fs = module.require("fs");
 
 module.exports.run = async (client, message, args) => {    
     let mcmember = await args[0];
-
-    if (!mcmember) {
-        return await message.channel.send({embed: {
-            color: 3447003,
-            title: "Mention a valid minecraft member!"
-        }}).then(msg => msg.delete(2000));
-    }
+    if (!mcmember) return message.delete(), message.channel.send("Mention a valid minecraft member!").then(msg => msg.delete(2000));
 
     request("https://use.gameapis.net/mc/player/profile/" + mcmember, function (error, response, body) {
-        if (response.statusCode == 400) {
-            message.delete();
-            return message.channel.send({embed: {
-                color: 3447003,
-                title: "Mention a valid minecraft member!"
-            }}).then(msg => msg.delete(2000));
-        }
-
+        if (response.statusCode == 400) return message.delete(), message.channel.send("Mention a valid minecraft member!").then(msg => msg.delete(2000));
         const mcdata = JSON.parse(body);
        
-        let profileName = mcdata.name; 
-        let profileID = mcdata.id;
-        let profileUUID = mcdata.uuid_formatted; 
-        let profileAvatar = "https://crafatar.com/avatars/" + profileUUID;
-        let profileBody = "https://crafatar.com/renders/body/" + profileUUID;
+        if (!mcdata.id) return message.delete(), message.channel.send("Mention a valid minecraft member!").then(msg => msg.delete(2000));
+        
         var profileSkin;
-
         if (mcdata.properties_decoded.textures.SKIN) {
             profileSkin = mcdata.properties_decoded.textures.SKIN.url;
         } else {
             profileSkin = "No Skin Found"
         }
        
-        message.channel.send({embed: {
-            color: 3447003,
-            author: {
-                name: profileName,
-                icon_url: profileAvatar
-            },
-            thumbnail: {
-                url: profileBody
-            },
-            fields: [
-            {
-                name: "Profile ID",
-                value: profileID
-            },
-            {
-                name: "UUID",
-                value: `${profileUUID}`
-            },
-            {
-                name: "Skin",
-                value: profileSkin,
-                url: profileSkin
-            },
-            ],
-                timestamp: new Date(),
-            }
-        });
+        let embed = new Discord.RichEmbed()
+          .setColor("4253f4")
+          .setAuthor(mcdata.name, "https://minotar.net/avatar/" + mcdata.name)
+          .setThumbnail("https://minotar.net/body/" + mcdata.name + "/100.png")
+          .addField("Profile ID", mcdata.id, true)
+          .addField("Premium Account", "Yes", true)
+          .addField("UUID", mcdata.uuid_formatted, true)
+          .addField("Skin", profileSkin, true)
+          .setTimestamp()
+      
+        message.channel.send(embed);
     });
 }
